refactor: move useTabsList to its own folder.
