+++ packages/mui-base/src/TabsListUnstyled/index.ts
@@ -3,6 +3,3 @@ export * from './TabsListUnstyled.types';
 
 export { default as tabsListUnstyledClasses } from './tabsListUnstyledClasses';
 export * from './tabsListUnstyledClasses';
-
-export { default as useTabsList } from './useTabsList';
-export * from './useTabsList.types';
