+++ packages/mui-base/src/index.js
@@ -22,6 +22,8 @@ export * from './useSwitch';
 
 export * from './useTabPanel';
 
+export * from './useTabsList';
+
 export { default as BadgeUnstyled } from './BadgeUnstyled';
 export * from './BadgeUnstyled';
