+++ packages/mui-base/src/useTabsList/index.ts
@@ -1,0 +1,5 @@
+export { default as useTabsList } from './useTabsList';
+export { default } from './useTabsList';
+export * from './useTabsList';
+
+export * from './useTabsList.types';
